Update panels to scene system
