+++ src/panel/panel_hover.c
@@ -5,16 +5,15 @@
 ** General foos for panels
 */
 
-#include "ui_panels/button.h"
-#include "program.h"
 #include "theme.h"
+#include "content.h"
 #include "tools.h"
 
 void on_panel_enter(panel_t *panel)
 {
     if (!panel)
         return;
-    if (is_concerned(panel->type)) {
+    if (is_interactable(panel->type)) {
         if (panel->shape)
             sfRectangleShape_setFillColor(panel->shape, BUTTON_HOVER);
     }
@@ -24,7 +23,7 @@ void on_panel_leave(panel_t *panel)
 {
     if (!panel)
         return;
-    if (is_concerned(panel->type)) {
+    if (is_interactable(panel->type)) {
         if (panel->shape)
             sfRectangleShape_setFillColor(panel->shape, BUTTON_BASE);
     }
@@ -34,7 +33,7 @@ void on_panel_pressed(panel_t *panel)
 {
     if (!panel)
         return;
-    if (is_concerned(panel->type)) {
+    if (is_interactable(panel->type)) {
         if (panel->shape)
             sfRectangleShape_setFillColor(panel->shape, BUTTON_PRESS);
     }
@@ -44,7 +43,7 @@ void on_panel_released(panel_t *panel)
 {
     if (!panel)
         return;
-    if (is_concerned(panel->type)) {
+    if (is_interactable(panel->type)) {
         if (panel->shape)
             sfRectangleShape_setFillColor(panel->shape, BUTTON_HOVER);
     }
