+++ src/panel/panel_select.c
@@ -10,11 +10,11 @@
 #include "theme.h"
 #include "tools.h"
 
-void on_panel_select(panel_t *panel, program_t *p)
+void on_panel_select(panel_t *panel)
 {
     if (!panel)
         return;
-    if (is_concerned(panel->type)) {
+    if (is_interactable(panel->type)) {
         if (panel->shape)
             sfRectangleShape_setFillColor(panel->shape, BUTTON_SELECT);
     }
@@ -26,7 +26,7 @@ void on_panel_select(panel_t *panel, program_t *p)
     if (panel->type == PTYPE_EVBUTT) {
         panel_button_t *bpanel = (panel_button_t *)panel->data;
         if (bpanel->on_select)
-            bpanel->on_select(p);
+            bpanel->on_select(bpanel->trgt);
     }
 }
 
@@ -34,13 +34,13 @@ void on_panel_unselect(panel_t *panel)
 {
     if (!panel)
         return;
-    if (is_concerned(panel->type)) {
+    if (is_interactable(panel->type)) {
         if (panel->shape)
             sfRectangleShape_setFillColor(panel->shape, BUTTON_BASE);
     }
     if (panel->type == PTYPE_DDBUTT) {
         panel_button_t *bpanel = (panel_button_t *)panel->data;
         if (bpanel->on_unselect)
-            bpanel->on_unselect(panel);
+            bpanel->on_unselect(bpanel->trgt);
     }
 }
