+++ src/tools.c
@@ -6,16 +6,6 @@
 */
 
 #include <stdlib.h>
-#include "panel/panel.h"
-
-int is_concerned(ptype_t t)
-{
-    if (t == PTYPE_DDBUTT ||
-        t == PTYPE_EVBUTT ||
-        t == PANEL_TYPE_INPUT)
-        return 1;
-    return 0;
-}
 
 int fast_atoi(const char *str)
 {
@@ -25,17 +15,6 @@ int fast_atoi(const char *str)
         str++;
     }
     return val;
-}
-
-void save_as_png(sfTexture* texture, const char* path)
-{
-    sfImage *image = sfTexture_copyToImage(texture);
-    if (!sfImage_saveToFile(image, path)) {
-        sfImage_destroy(image);
-        return;
-    }
-    sfImage_destroy(image);
-    return;
 }
 
 char *str(const char *str, int i)
